[Shared] Graph Composable now properly maps raw X data values to x canvas coordinates rather than use data indices as X values
- previously, using data indices as X values was a quick way
  to implement line rendering

+++ app/src/main/java/com/zackjp/devicedx/feature/traffic/util/TrafficGraphUtil.kt
@@ -43,11 +43,11 @@ class TrafficGraphUtil @Inject constructor() {
             val currentSecData = bucketedDataBySecond[currentSec]
             val priorSecData = bucketedDataBySecond[priorSec]
             if (currentSecData == null || priorSecData == null) {
-                TrafficMetric(currentSec, 0f)
+                TrafficMetric(currentSec, 0L)
             } else {
                 TrafficMetric(
                     currentSec,
-                    (currentSecData.rxBytes - priorSecData.rxBytes).coerceAtLeast(0).toFloat()
+                    (currentSecData.rxBytes - priorSecData.rxBytes).coerceAtLeast(0)
                 )
             }
         }
